CK9n: Use `enable` instead of `start` for Neural Networks service
* The MTK Neural Networks service is started via a property trigger.
This HAL starts a lot earlier than other HALs since it is started
with "start" instead of "enable", the difference between them being
that "start" will immediately cause the service to try and start, while
"enable" effectively removes the "disabled" attribute from the service
(so in the case of a HAL, it will be started with all the other services
in class hal instead of immediately).

* This initially didn't have any side effects until I noticed the Neural Networks
service constantly crashing in offline charger mode. To fix this, simply
convert the "start" command to an "enable" to ensure that this service is not
only started at the right time (in class hal), but also only started when
needed (and not in offline charger mode).